requirements.txt
Flask==2.0.3
werkzeug==2.3.7
tensorflow==2.18.0
Pillow
flask-cors
gunicorn==20.1.0  # For deployment on Render
